1.0 Release

+++ WindowsFormsApplication1/WindowsFormsApplication1/frmMain.cs
@@ -32,15 +32,21 @@ namespace WindowsFormsApplication1
 
         private void button3_Click(object sender, EventArgs e)
         {
-            //cycles through each selected row, deleting based on the id tag
-            int selectedrowindex = dgPassword.SelectedCells[0].RowIndex;
-            DataGridViewRow selectedRow = dgPassword.Rows[selectedrowindex];
-            string id = selectedRow.Cells[0].Value.ToString();
-            Console.WriteLine(id);
-            DBManager.DeleteId(id);
-            
-            //refreshes datagrid view
-            this.refreshDG();
+            //selects the first value which is id
+            try {
+                int selectedrowindex = dgPassword.SelectedCells[0].RowIndex;
+                DataGridViewRow selectedRow = dgPassword.Rows[selectedrowindex];
+                string id = selectedRow.Cells[0].Value.ToString();
+
+                if (MessageBox.Show("Are you sure you want to delete?", "Confirm delete", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                    DBManager.DeleteId(id);
+                //refreshes datagrid view
+                this.refreshDG();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("Please select a valid record to delete!", "Error");
+            }
         }
 
 
@@ -60,15 +66,34 @@ namespace WindowsFormsApplication1
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
+
+            try { 
             int selectedrowindex = dgPassword.SelectedCells[0].RowIndex;
-            DataGridViewRow selectedRow = dgPassword.Rows[selectedrowindex];
+            DataGridViewRow selectedRow = dgPassword.Rows[selectedrowindex];   
             string id = selectedRow.Cells[0].Value.ToString();
             string username = selectedRow.Cells[1].Value.ToString();
             string password = selectedRow.Cells[2].Value.ToString();
             string website = selectedRow.Cells[3].Value.ToString();
-
             frmEdit temp = new frmEdit(this, id, username, password, website);
             temp.Show();
+            }
+
+            catch(Exception ex)
+            {
+                MessageBox.Show("Please select a valid record to edit!", "Error");
+            }
+            
+        }
+
+        private void exitToolStripMenuItem1_Click(object sender, EventArgs e)
+        {
+            Application.Exit();
+        }
+
+        private void aboutToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            aboutBox about = new aboutBox();
+            about.Show();
         }
     }
     }
